update md-02 assignment.
