+++ assignments/md-02/am-02-1-data-visualisation.qmd
@@ -4,46 +4,74 @@ title: "Data visualisation with ggplot2"
 
 ::: {.callout-important}
 ## Have you completed the assignments for module 1?
-A pre-requisite for this homework is that you worked through the "Interaction with GitHub" assignment of module 1. You will need your GitHub Personal Access Token to push your committed changes back to GitHub. 
+A pre-requisite for this homework is that you worked through the "Interaction with GitHub" assignment of module 1. You will need your GitHub Personal Access Token to push your committed changes back to GitHub.
+:::
+
+::: {.callout-note}
+## Reduced level of detail
+As we progress through the course, we're gradually reducing the level of instructional detail. This assignment follows the same steps as module 1, but without screenshots. If you need visual guidance, refer back to the module 1 assignments.
 :::
 
 ## Step 1: Clone your module 2 assignments repository
 
-1. Open github.com in your browser and navigate to the GitHub organisation for the course (or use your bookmark).
-2. Find the repository md-02-assignments-USERNAME that ends with your GitHub username, and open it.
-3. Click on the green "Code" button.
-4. Copy the HTTPS URL to your clipboard.
-5. Open the ds4owd workspace on [posit.cloud](https://posit.cloud/spaces/426916/content/all?sort=name_asc)
-6. Click New Project > New Project from Git Repository
-7. Paste the HTTPS URL from GitHub into the "URL of your Git Repository" field.
-8. Wait until the project is deployed.
-9. Open `md-02a-data-visualization.qmd` and follow the instructions to work through the tasks.
+1. Open <https://github.com/> in your browser.
+2. Navigate to the GitHub organization for the course: [{{< var course.github-org >}}]({{< var course.github-org >}})
+3. Find the repository `md-02-assignments-USERNAME` that [ends with your GitHub username]{.highlight-yellow}, and open it by clicking on the repository name.
+   - Replace `USERNAME` with your actual GitHub username.
+   - For example, if your username is `johnsmith`, the repository will be `md-02-assignments-johnsmith`.
+4. Click on the green [Code]{.highlight-yellow} button.
+5. Copy the HTTPS URL to your clipboard by clicking on the clipboard icon next to the URL.
+6. Open the [Content]{.highlight-yellow} tab of the ds4owd-002 workspace on Posit Cloud: [{{< var platforms.posit-cloud >}}/content/all?sort=name_asc]({{< var platforms.posit-cloud >}}/content/all?sort=name_asc)
+7. Click the blue button [New Project]{.highlight-yellow} > [New Project from Git Repository]{.highlight-yellow}
+8. Paste the HTTPS URL from GitHub into the [URL of your Git Repository]{.highlight-yellow} field and click [OK]{.highlight-yellow}. Leave the checkbox under Packages checked.
+9. Wait until the project is deployed. This may take a few minutes, depending on your internet connection.
+10. Open `md-02a-data-visualization.qmd` and follow the instructions to work through the tasks.
 
-## Step 2: Commit and push your changes 
+## Step 2: Commit and push your changes
 
 The below is also described as Task 14 in `md-02a-data-visualization.qmd`
 
 1. Navigate to the Git pane in the top-right window of RStudio
-2. Check the box next to all files to stage then it for a commit
-3. Click on the "Commit" button
-4. Enter a commit message in the "Commit message" field
-5. Click on the "Commit" button
-6. Click on the "Push" button
-7. Enter your GitHub username and GitHub Personal Access Token (PAT) in the "Username" and "Password" fields
+2. Check the box next to all files to stage them for a commit
+3. Click on the [Commit]{.highlight-yellow} button. A new window opens.
+4. Enter a commit message in the [Commit message]{.highlight-yellow} field in the top right corner. For example: completed data visualization tasks
+5. Click on the [Commit]{.highlight-yellow} button
+6. If you see a message that confirms your commit, click on [Close]{.highlight-yellow}. If you do not see this message, make sure you have staged the files (checked the box next to them) for commit and entered a commit message.
+7. Click on the [Push]{.highlight-yellow} button in the top right corner.
+8. Enter your GitHub username and click OK
+9. Enter your GitHub Personal Access Token (PAT) in the field and click OK. This is the personal access token you created in Module 1.
 
-::: callout-important
+If you see a message that says `HEAD -> main`, then you have successfully pushed your changes to the remote repository on GitHub. Click Close. [If you do not see this message]{.highlight-yellow}, make sure you have entered your GitHub username and GitHub PAT correctly.
+
+::: {.callout-warning}
+## PAT Storage Warning
+
+If you have stored your PAT in a Word file, a blank space might have been copied along with the PAT. We recommend using a TXT file (Notepad on Windows or TextEdit on Mac) or a password manager to store your PAT securely.
+:::
+
+::: {.callout-important}
 ## Do not use your GitHub password
 
-You need to enter the GitHub Personal Access Token (PAT) you created in Step 1 to push your changes back to GitHub.
+You need to enter the GitHub Personal Access Token (PAT) you created in Module 1 to push your changes back to GitHub.
 :::
 
 ## Step 3: Open an issue on GitHub
 
-1. Open [github.com](https://github.com/) in your browser.
-2. Navigate to the GitHub organisation for the course.
-3. Find the repository md-02-assignments that ends with your GitHub username.
-4. Click on the "Issues" tab.
-5. Click on the green "New issue" button.
-6. In the "Title" field write: "Completed module 1 assignments".
-7. In the "Leave a comment" field, tag the course instructors `@seawaR` `@massarin` `@larnsce`
+1. Open <https://github.com/> in your browser.
+2. Navigate to the GitHub organization for the course: [{{< var course.github-org >}}]({{< var course.github-org >}})
+3. Find the repository `md-02-assignments-USERNAME` that [ends with your GitHub username]{.highlight-yellow}, and open it by clicking on the repository name.
+   - Replace `USERNAME` with your actual GitHub username.
+   - For example, if your username is `johnsmith`, the repository will be `md-02-assignments-johnsmith`.
+
+::: {.callout-tip}
+You can search for your repository by typing your username in the search bar just below the Repository heading.
+:::
+
+4. You can verify here that your changes have been pushed to the remote repository by looking at the text next to the `md-02a-data-visualization.qmd` file. It should display the commit message you entered in the previous step.
+5. Click on the [Issues]{.highlight-yellow} tab.
+6. Click on the green [New issue]{.highlight-yellow} button.
+7. In the [Title]{.highlight-yellow} field write: [Completed module 2 assignment]{.highlight-yellow}.
+8. In the [Leave a comment]{.highlight-yellow} field, tag the course instructors `@seawaR` `@massarin` `@larnsce` and [add a question about this module]{.highlight-yellow}. We would like to see [one question from every participant]{.highlight-yellow}. 
+9. Scroll down the page a bit and click the green [Submit new issue]{.highlight-yellow} button.
+
