+++ docs/sitemap.xml
@@ -22,7 +22,7 @@
   </url>
   <url>
     <loc>https://ds4owd-002.github.io/website/modules/md-01.html</loc>
-    <lastmod>2025-09-11T11:50:38.835Z</lastmod>
+    <lastmod>2025-09-17T06:26:33.843Z</lastmod>
   </url>
   <url>
     <loc>https://ds4owd-002.github.io/website/modules/md-02.html</loc>
@@ -82,7 +82,7 @@
   </url>
   <url>
     <loc>https://ds4owd-002.github.io/website/assignments/md-02/am-02-1-data-visualisation.html</loc>
-    <lastmod>2025-09-10T19:18:35.293Z</lastmod>
+    <lastmod>2025-09-17T18:08:40.035Z</lastmod>
   </url>
   <url>
     <loc>https://ds4owd-002.github.io/website/assignments/md-03/am-03-1-data-transformation.html</loc>
@@ -130,7 +130,7 @@
   </url>
   <url>
     <loc>https://ds4owd-002.github.io/website/slides/lec-02-visualisation.html</loc>
-    <lastmod>2025-09-10T17:49:02.194Z</lastmod>
+    <lastmod>2025-09-17T06:29:25.791Z</lastmod>
   </url>
   <url>
     <loc>https://ds4owd-002.github.io/website/slides/lec-04-tidy-data.html</loc>
