+++ slides/lec-02-visualisation_files/figure-revealjs/barplot-penguins-1.png
