+++ slides/lec-02-visualisation_files/figure-revealjs/histogram-penguins-1.png
