+++ slides/lec-02-visualisation_files/figure-revealjs/scatterplot-gdp-lifeExp-1.png
